Move SVG string fixups/loading into separate file
The VM uses the SvgRenderer solely to convert SVGs from version 2.
This allows the VM to simply import loadSvgString without bringing in
the entire SvgRenderer along with it. The next commit will move toString
to a separate file as well, so the VM's dependency on SvgRenderer can be
entirely removed.
